backend/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
python-dotenv
celery
redis
sqlalchemy
asyncpg
psycopg2-binary